requirements.txt
appdirs==1.4.4
APScheduler==3.9.1
asgiref==3.5.2
astroid==2.11.7
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
chromedriver==2.24.1
colorama==0.4.5
cryptography==37.0.4
dill==0.3.5.1
Django==4.0.5
django-bootstrap-v5==1.0.11
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
h11==0.13.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
lxml==4.9.1
mccabe==0.7.0
numpy==1.23.1
outcome==1.2.0
pandas==1.4.3
phonenumbers==8.12.50
platformdirs==2.5.2
psycopg2==2.9.3
pycparser==2.21
pylint==2.14.4
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.1
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
wrapt==1.14.1
wsproto==1.1.0
